requirements.txt
amas==1.0
biopython==1.85
certifi==2025.8.3
charset-normalizer==3.4.3
dotenv==0.9.9
idna==3.10
numpy==2.3.3
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
